backend/requirements.pip.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
python-dotenv
djangorestframework-simplejwt==5.3.1
djoser==2.2.0
django-cors-headers